+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+2023.04.26
+- Added the `-maf` option that allows the user to specify a minor allele frequency cutoff for ingroups in the MK tests; alleles below the specified frequency will be excluded. The default `-maf` cutoff is 1 / 2N where N is the number of ingroup samples.
+- Added `--no-fixed-ingroup` option that will exclude sites from MK tests if all ingroup samples are fixed for an alternate allele, which is likely an error in the reference
+
 2023.04.20
 - Converted error for transcripts with exons on differing strands to a warning, and added warning for transcripts with no coding exons